Moved comments-ui unit tests from src/ to test/unit/
Aligns comments-ui test organization with other apps in the monorepo
(portal, admin-x-settings, posts, stats) which keep unit tests in a
separate test/unit/ directory that mirrors the src/ structure. This
makes it easier to find tests and maintains consistency across apps.
